requirements.txt
asgiref==3.8.1
attrs==25.3.0
chardet==5.2.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
fire==0.7.0
fonttools==4.56.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
numpy==2.2.4
opencv-python-headless==4.11.0.86
pdf2docx==0.5.8
pillow==11.1.0
PyJWT==2.9.0
PyMuPDF==1.25.4
PyPDF2==3.0.1
python-docx==1.1.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.3.1
rpds-py==0.24.0
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
